refactor(DataNotExistException.java): Entity 조회 시, 존재하지 않는 Entity 에 대해 DataNotExistException 적용
Entity 조회 시, 존재하지 않는 Entity 에 대해 DataNotExistException 을 적용한다.
- DataNotExistException : NotExistLineException -> DataNotExistException 로 변경
