+++ src/main/java/nextstep/subway/exception/DataNotExistException.java
@@ -1,0 +1,11 @@
+package nextstep.subway.exception;
+
+public class DataNotExistException extends RuntimeException {
+
+    public DataNotExistException() {
+    }
+
+    public DataNotExistException(String message) {
+        super(message);
+    }
+}
