--- src/main/java/nextstep/subway/line/application/NotExistLineException.java
@@ -1,11 +1,0 @@
-package nextstep.subway.line.application;
-
-public class NotExistLineException extends RuntimeException {
-
-    public NotExistLineException() {
-    }
-
-    public NotExistLineException(String message) {
-        super(message);
-    }
-}
